Refactor structure and clean imports
