+++ lib/shared/widgets/custom_search_bar.dart
